requirements.txt
rich
psutil
pypresence
discord.py-self
colorama
inquirer
questionary
requests
readchar
pyfiglet
asyncio
aiohttp
PyQt6
pygame
